Adds auto decompression if mothur has trouble reading gz files in make.contigs
#634

+++ source/commands/aligncommand.cpp
@@ -519,7 +519,7 @@ long long AlignCommand::createProcesses(string alignFileName, string reportFileN
 	try {
         
         vector<linePair> lines;
-        vector<unsigned long long> positions;
+        vector<double> positions;
 #if defined NON_WINDOWS
         positions = util.divideFile(filename, processors);
         for (int i = 0; i < (positions.size()-1); i++) {	lines.push_back(linePair(positions[i], positions[(i+1)]));	}
