+++ source/commands/degapseqscommand.cpp
@@ -255,7 +255,7 @@ long long DegapSeqsCommand::createProcesses(string filename, string outputFileNa
         vector<linePair> lines;
         
         long long num = 0;
-        vector<unsigned long long> positions;
+        vector<double> positions;
 #if defined NON_WINDOWS
         positions = util.divideFile(filename, processors);
         for (int i = 0; i < (positions.size()-1); i++) {	lines.push_back(linePair(positions[i], positions[(i+1)]));	}
